+++ libs/package.json
@@ -61,7 +61,7 @@
     "@rollup/plugin-node-resolve": "13.1.3",
     "@rollup/plugin-typescript": "8.3.1",
     "@tsconfig/node16-strictest": "1.0.0",
-    "@types/async-retry": "^1.4.3",
+    "@types/async-retry": "1.4.3",
     "@types/bs58": "^4.0.1",
     "@types/expect": "24.3.0",
     "@types/hashids": "2.0.1",
