+++ libs/src/index.js
@@ -12,7 +12,7 @@ const { CreatorNode } = require('./services/creatorNode')
 const DiscoveryProvider = require('./services/discoveryProvider/index')
 const Wormhole = require('./services/wormhole')
 const { AudiusABIDecoder } = require('./services/ABIDecoder')
-const SchemaValidator = require('./services/schemaValidator')
+const { SchemaValidator } = require('./services/schemaValidator')
 const { UserStateManager } = require('./userStateManager')
 const SanityChecks = require('./sanityChecks')
 const { Utils, Captcha } = require('./utils')
